Check deltatime dimensionality
This closes #23.

The deltatime array for some products has one dimension, i.e. one
timestamp for each scan line.

For other products, however, the array has the same dimensionality as
the data, i.e. a 2d-array with the timestamps for each scan line
repeated for every pixel in that line.

This patch fixes this problem by checking the dimensionality of the
deltatime array.